+++ example/lib/main.dart
@@ -16,7 +16,8 @@ void main() {
             color: Color.fromRGBO(30, 60, 87, 1),
           ),
         ),
-        body: const FractionallySizedBox(widthFactor: 1, child: PinputExample()),
+        body:
+            const FractionallySizedBox(widthFactor: 1, child: PinputExample()),
       ),
     ),
   );
@@ -62,6 +63,31 @@ class _PinputExampleState extends State<PinputExample> {
       ),
     );
 
+    // return TextField(
+    //   contextMenuBuilder: (_, EditableTextState editableTextState) {
+    //     print('HEHE');
+    //     return AdaptiveTextSelectionToolbar(
+    //       anchors: editableTextState.contextMenuAnchors,
+    //       children: editableTextState.contextMenuButtonItems.map((ContextMenuButtonItem buttonItem) {
+    //         return CupertinoButton(
+    //           borderRadius: null,
+    //           color: const Color(0xffaaaa00),
+    //           disabledColor: const Color(0xffaaaaff),
+    //           onPressed: buttonItem.onPressed,
+    //           padding: const EdgeInsets.all(10.0),
+    //           pressedOpacity: 0.7,
+    //           child: SizedBox(
+    //             width: 200.0,
+    //             child: Text(
+    //               CupertinoTextSelectionToolbarButton.getButtonLabel(context, buttonItem),
+    //             ),
+    //           ),
+    //         );
+    //       }).toList(),
+    //     );
+    //   },
+    // );
+
     /// Optionally you can use form to validate the Pinput
     return Form(
       key: formKey,
@@ -74,7 +100,8 @@ class _PinputExampleState extends State<PinputExample> {
             child: Pinput(
               controller: pinController,
               focusNode: focusNode,
-              androidSmsAutofillMethod: AndroidSmsAutofillMethod.smsUserConsentApi,
+              androidSmsAutofillMethod:
+                  AndroidSmsAutofillMethod.smsUserConsentApi,
               listenForMultipleSmsOnAndroid: true,
               defaultPinTheme: defaultPinTheme,
               validator: (value) {