+++ lib/src/pinput.dart
@@ -86,7 +86,6 @@ class Pinput extends StatefulWidget {
     this.closeKeyboardWhenCompleted = true,
     this.keyboardType = TextInputType.number,
     this.textCapitalization = TextCapitalization.none,
-    this.toolbarOptions = const ToolbarOptions(paste: true),
     this.slideTransitionBeginOffset,
     this.cursor,
     this.keyboardAppearance,
@@ -107,7 +106,8 @@ class Pinput extends StatefulWidget {
     this.errorTextStyle,
     this.pinputAutovalidateMode = PinputAutovalidateMode.onSubmit,
     this.scrollPadding = const EdgeInsets.all(20),
-    this.initialValue = _defaultContextMenuBuilder,
+    this.contextMenuBuilder = _defaultContextMenuBuilder,
+    this.onTapOutside,
     Key? key,
   })  : assert(obscuringCharacter.length == 1),
         assert(length > 0),
@@ -287,13 +287,6 @@ class Pinput extends StatefulWidget {
   /// [TextInputType.multiline] and [TextInputAction.done] otherwise.
   final TextInputAction? textInputAction;
 
-  /// Configuration of toolbar options.
-  ///
-  /// If not set, select all and paste will default to be enabled. Copy and cut
-  /// will be disabled if [obscureText] is true. If [readOnly] is true,
-  /// paste and cut will be disabled regardless.
-  final ToolbarOptions toolbarOptions;
-
   /// See [EditableText.autofillHints]
   final Iterable<String>? autofillHints;
 
@@ -355,7 +348,18 @@ class Pinput extends StatefulWidget {
   ///  * [AdaptiveTextSelectionToolbar], which is built by default.
   final EditableTextContextMenuBuilder? contextMenuBuilder;
 
-  static Widget _defaultContextMenuBuilder(BuildContext context, EditableTextState editableTextState) {
+  /// A callback to be invoked when a tap is detected outside of this [TapRegion]
+  /// The [PointerDownEvent] passed to the function is the event that caused the
+  /// notification. If this region is part of a group
+  /// then it's possible that the event may be outside of this immediate region,
+  /// although it will be within the region of one of the group members.
+  /// This is useful if you want to unfocus the [Pinput] when user taps outside of it
+  final TapRegionCallback? onTapOutside;
+
+  static Widget _defaultContextMenuBuilder(
+    BuildContext context,
+    EditableTextState editableTextState,
+  ) {
     return AdaptiveTextSelectionToolbar.editableText(
       editableTextState: editableTextState,
     );
@@ -367,61 +371,364 @@ class Pinput extends StatefulWidget {
   @override
   void debugFillProperties(DiagnosticPropertiesBuilder properties) {
     super.debugFillProperties(properties);
-    properties.add(DiagnosticsProperty<PinTheme>('defaultPinTheme', defaultPinTheme, defaultValue: null));
-    properties.add(DiagnosticsProperty<PinTheme>('focusedPinTheme', focusedPinTheme, defaultValue: null));
-    properties.add(DiagnosticsProperty<PinTheme>('submittedPinTheme', submittedPinTheme, defaultValue: null));
-    properties.add(DiagnosticsProperty<PinTheme>('followingPinTheme', followingPinTheme, defaultValue: null));
-    properties.add(DiagnosticsProperty<PinTheme>('disabledPinTheme', disabledPinTheme, defaultValue: null));
-    properties.add(DiagnosticsProperty<PinTheme>('errorPinTheme', errorPinTheme, defaultValue: null));
-    properties.add(DiagnosticsProperty<TextEditingController>('controller', controller, defaultValue: null));
-    properties.add(DiagnosticsProperty<FocusNode>('focusNode', focusNode, defaultValue: null));
-    properties.add(DiagnosticsProperty<bool>('enabled', enabled, defaultValue: true));
-    properties.add(DiagnosticsProperty<bool>('closeKeyboardWhenCompleted', closeKeyboardWhenCompleted, defaultValue: true));
-    properties.add(DiagnosticsProperty<TextInputType>('keyboardType', keyboardType, defaultValue: TextInputType.number));
-    properties.add(DiagnosticsProperty<int>('length', length, defaultValue: PinputConstants._defaultLength));
-    properties.add(DiagnosticsProperty<ValueChanged<String>?>('onCompleted', onCompleted, defaultValue: null));
-    properties.add(DiagnosticsProperty<ValueChanged<String>?>('onChanged', onChanged, defaultValue: null));
-    properties.add(DiagnosticsProperty<ValueChanged<String>?>('onClipboardFound', onClipboardFound, defaultValue: null));
-    properties.add(DiagnosticsProperty<VoidCallback?>('onTap', onTap, defaultValue: null));
-    properties.add(DiagnosticsProperty<VoidCallback?>('onLongPress', onLongPress, defaultValue: null));
-    properties.add(DiagnosticsProperty<Widget?>('preFilledWidget', preFilledWidget, defaultValue: null));
-    properties.add(DiagnosticsProperty<Widget?>('cursor', cursor, defaultValue: null));
-    properties.add(DiagnosticsProperty<Widget?>('separator', separator, defaultValue: PinputConstants._defaultSeparator));
-    properties.add(DiagnosticsProperty<Widget?>('obscuringWidget', obscuringWidget, defaultValue: null));
-    properties.add(DiagnosticsProperty<List<int>?>('separatorPositions', separatorPositions, defaultValue: null));
-    properties.add(DiagnosticsProperty<MainAxisAlignment>('mainAxisAlignment', mainAxisAlignment, defaultValue: MainAxisAlignment.center));
-    properties.add(DiagnosticsProperty<AlignmentGeometry>('pinContentAlignment', pinContentAlignment, defaultValue: Alignment.center));
-    properties.add(DiagnosticsProperty<Curve>('animationCurve', animationCurve, defaultValue: Curves.easeIn));
-    properties.add(DiagnosticsProperty<Duration>('animationDuration', animationDuration, defaultValue: PinputConstants._animationDuration));
-    properties.add(DiagnosticsProperty<PinAnimationType>('pinAnimationType', pinAnimationType, defaultValue: PinAnimationType.scale));
-    properties.add(DiagnosticsProperty<Offset?>('slideTransitionBeginOffset', slideTransitionBeginOffset, defaultValue: null));
-    properties.add(DiagnosticsProperty<bool>('enabled', enabled, defaultValue: true));
-    properties.add(DiagnosticsProperty<bool>('readOnly', readOnly, defaultValue: false));
-    properties.add(DiagnosticsProperty<bool>('obscureText', obscureText, defaultValue: false));
-    properties.add(DiagnosticsProperty<bool>('autofocus', autofocus, defaultValue: false));
-    properties.add(DiagnosticsProperty<bool>('useNativeKeyboard', useNativeKeyboard, defaultValue: false));
-    properties.add(DiagnosticsProperty<bool>('toolbarEnabled', toolbarEnabled, defaultValue: true));
-    properties.add(DiagnosticsProperty<bool>('showCursor', showCursor, defaultValue: true));
-    properties.add(DiagnosticsProperty<String>('obscuringCharacter', obscuringCharacter, defaultValue: '•'));
-    properties.add(DiagnosticsProperty<bool>('obscureText', obscureText, defaultValue: false));
-    properties.add(DiagnosticsProperty<bool>('enableSuggestions', enableSuggestions, defaultValue: true));
-    properties.add(DiagnosticsProperty<List<TextInputFormatter>>('inputFormatters', inputFormatters, defaultValue: const <TextInputFormatter>[]));
-    properties.add(EnumProperty<TextInputAction>('textInputAction', textInputAction, defaultValue: TextInputAction.done));
-    properties.add(EnumProperty<TextCapitalization>('textCapitalization', textCapitalization, defaultValue: TextCapitalization.none));
-    properties.add(DiagnosticsProperty<Brightness>('keyboardAppearance', keyboardAppearance, defaultValue: null));
-    properties.add(DiagnosticsProperty<TextInputType>('keyboardType', keyboardType, defaultValue: TextInputType.number));
-    properties.add(DiagnosticsProperty<ToolbarOptions>('toolbarOptions', toolbarOptions, defaultValue: const ToolbarOptions(paste: true)));
-    properties.add(DiagnosticsProperty<Iterable<String>?>('autofillHints', autofillHints, defaultValue: null));
-    properties.add(DiagnosticsProperty<TextSelectionControls?>('selectionControls', selectionControls, defaultValue: null));
-    properties.add(DiagnosticsProperty<String?>('restorationId', restorationId, defaultValue: null));
-    properties.add(DiagnosticsProperty<AppPrivateCommandCallback?>('onAppPrivateCommand', onAppPrivateCommand, defaultValue: null));
-    properties.add(DiagnosticsProperty<MouseCursor?>('mouseCursor', mouseCursor, defaultValue: null));
-    properties.add(DiagnosticsProperty<TextStyle?>('errorTextStyle', errorTextStyle, defaultValue: null));
-    properties.add(DiagnosticsProperty<PinputErrorBuilder?>('errorBuilder', errorBuilder, defaultValue: null));
-    properties.add(DiagnosticsProperty<FormFieldValidator<String>?>('validator', validator, defaultValue: null));
-    properties.add(DiagnosticsProperty<PinputAutovalidateMode>('pinputAutovalidateMode', pinputAutovalidateMode, defaultValue: PinputAutovalidateMode.onSubmit));
-    properties.add(DiagnosticsProperty<HapticFeedbackType>('hapticFeedbackType', hapticFeedbackType, defaultValue: HapticFeedbackType.disabled));
-    properties.add(DiagnosticsProperty<String?>('senderPhoneNumber', senderPhoneNumber, defaultValue: null));
-    properties.add(DiagnosticsProperty<EditableTextContextMenuBuilder?>('contextMenuBuilder', contextMenuBuilder, defaultValue: _defaultContextMenuBuilder));
+    properties.add(
+      DiagnosticsProperty<PinTheme>(
+        'defaultPinTheme',
+        defaultPinTheme,
+        defaultValue: null,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<PinTheme>(
+        'focusedPinTheme',
+        focusedPinTheme,
+        defaultValue: null,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<PinTheme>(
+        'submittedPinTheme',
+        submittedPinTheme,
+        defaultValue: null,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<PinTheme>(
+        'followingPinTheme',
+        followingPinTheme,
+        defaultValue: null,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<PinTheme>(
+        'disabledPinTheme',
+        disabledPinTheme,
+        defaultValue: null,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<PinTheme>(
+        'errorPinTheme',
+        errorPinTheme,
+        defaultValue: null,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<TextEditingController>(
+        'controller',
+        controller,
+        defaultValue: null,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<FocusNode>(
+        'focusNode',
+        focusNode,
+        defaultValue: null,
+      ),
+    );
+    properties
+        .add(DiagnosticsProperty<bool>('enabled', enabled, defaultValue: true));
+    properties.add(
+      DiagnosticsProperty<bool>(
+        'closeKeyboardWhenCompleted',
+        closeKeyboardWhenCompleted,
+        defaultValue: true,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<TextInputType>(
+        'keyboardType',
+        keyboardType,
+        defaultValue: TextInputType.number,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<int>(
+        'length',
+        length,
+        defaultValue: PinputConstants._defaultLength,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<ValueChanged<String>?>(
+        'onCompleted',
+        onCompleted,
+        defaultValue: null,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<ValueChanged<String>?>(
+        'onChanged',
+        onChanged,
+        defaultValue: null,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<ValueChanged<String>?>(
+        'onClipboardFound',
+        onClipboardFound,
+        defaultValue: null,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<VoidCallback?>('onTap', onTap, defaultValue: null),
+    );
+    properties.add(
+      DiagnosticsProperty<VoidCallback?>(
+        'onLongPress',
+        onLongPress,
+        defaultValue: null,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<Widget?>(
+        'preFilledWidget',
+        preFilledWidget,
+        defaultValue: null,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<Widget?>('cursor', cursor, defaultValue: null),
+    );
+    properties.add(
+      DiagnosticsProperty<Widget?>(
+        'separator',
+        separator,
+        defaultValue: PinputConstants._defaultSeparator,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<Widget?>(
+        'obscuringWidget',
+        obscuringWidget,
+        defaultValue: null,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<List<int>?>(
+        'separatorPositions',
+        separatorPositions,
+        defaultValue: null,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<MainAxisAlignment>(
+        'mainAxisAlignment',
+        mainAxisAlignment,
+        defaultValue: MainAxisAlignment.center,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<AlignmentGeometry>(
+        'pinContentAlignment',
+        pinContentAlignment,
+        defaultValue: Alignment.center,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<Curve>(
+        'animationCurve',
+        animationCurve,
+        defaultValue: Curves.easeIn,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<Duration>(
+        'animationDuration',
+        animationDuration,
+        defaultValue: PinputConstants._animationDuration,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<PinAnimationType>(
+        'pinAnimationType',
+        pinAnimationType,
+        defaultValue: PinAnimationType.scale,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<Offset?>(
+        'slideTransitionBeginOffset',
+        slideTransitionBeginOffset,
+        defaultValue: null,
+      ),
+    );
+    properties
+        .add(DiagnosticsProperty<bool>('enabled', enabled, defaultValue: true));
+    properties.add(
+      DiagnosticsProperty<bool>('readOnly', readOnly, defaultValue: false),
+    );
+    properties.add(
+      DiagnosticsProperty<bool>(
+        'obscureText',
+        obscureText,
+        defaultValue: false,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<bool>('autofocus', autofocus, defaultValue: false),
+    );
+    properties.add(
+      DiagnosticsProperty<bool>(
+        'useNativeKeyboard',
+        useNativeKeyboard,
+        defaultValue: false,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<bool>(
+        'toolbarEnabled',
+        toolbarEnabled,
+        defaultValue: true,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<bool>(
+        'showCursor',
+        showCursor,
+        defaultValue: true,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<String>(
+        'obscuringCharacter',
+        obscuringCharacter,
+        defaultValue: '•',
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<bool>(
+        'obscureText',
+        obscureText,
+        defaultValue: false,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<bool>(
+        'enableSuggestions',
+        enableSuggestions,
+        defaultValue: true,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<List<TextInputFormatter>>(
+        'inputFormatters',
+        inputFormatters,
+        defaultValue: const <TextInputFormatter>[],
+      ),
+    );
+    properties.add(
+      EnumProperty<TextInputAction>(
+        'textInputAction',
+        textInputAction,
+        defaultValue: TextInputAction.done,
+      ),
+    );
+    properties.add(
+      EnumProperty<TextCapitalization>(
+        'textCapitalization',
+        textCapitalization,
+        defaultValue: TextCapitalization.none,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<Brightness>(
+        'keyboardAppearance',
+        keyboardAppearance,
+        defaultValue: null,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<TextInputType>(
+        'keyboardType',
+        keyboardType,
+        defaultValue: TextInputType.number,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<Iterable<String>?>(
+        'autofillHints',
+        autofillHints,
+        defaultValue: null,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<TextSelectionControls?>(
+        'selectionControls',
+        selectionControls,
+        defaultValue: null,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<String?>(
+        'restorationId',
+        restorationId,
+        defaultValue: null,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<AppPrivateCommandCallback?>(
+        'onAppPrivateCommand',
+        onAppPrivateCommand,
+        defaultValue: null,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<MouseCursor?>(
+        'mouseCursor',
+        mouseCursor,
+        defaultValue: null,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<TextStyle?>(
+        'errorTextStyle',
+        errorTextStyle,
+        defaultValue: null,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<PinputErrorBuilder?>(
+        'errorBuilder',
+        errorBuilder,
+        defaultValue: null,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<FormFieldValidator<String>?>(
+        'validator',
+        validator,
+        defaultValue: null,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<PinputAutovalidateMode>(
+        'pinputAutovalidateMode',
+        pinputAutovalidateMode,
+        defaultValue: PinputAutovalidateMode.onSubmit,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<HapticFeedbackType>(
+        'hapticFeedbackType',
+        hapticFeedbackType,
+        defaultValue: HapticFeedbackType.disabled,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<String?>(
+        'senderPhoneNumber',
+        senderPhoneNumber,
+        defaultValue: null,
+      ),
+    );
+    properties.add(
+      DiagnosticsProperty<EditableTextContextMenuBuilder?>(
+        'contextMenuBuilder',
+        contextMenuBuilder,
+        defaultValue: _defaultContextMenuBuilder,
+      ),
+    );
   }
 }