+++ lib/src/pinput_state.dart
@@ -15,10 +15,11 @@ class _PinputState extends State<Pinput>
   late bool forcePressEnabled;
 
   @override
-  final GlobalKey<EditableTextState> editableTextKey = GlobalKey<EditableTextState>();
+  final GlobalKey<EditableTextState> editableTextKey =
+      GlobalKey<EditableTextState>();
 
   @override
-  bool get selectionEnabled => false;
+  bool get selectionEnabled => widget.toolbarEnabled;
 
   @override
   String get autofillId => _editableText!.autofillId;
@@ -37,7 +38,8 @@ class _PinputState extends State<Pinput>
   String? get _errorText => widget.errorText ?? _validatorErrorText;
 
   bool get _canRequestFocus {
-    final NavigationMode mode = MediaQuery.maybeOf(context)?.navigationMode ?? NavigationMode.traditional;
+    final NavigationMode mode = MediaQuery.maybeOf(context)?.navigationMode ??
+        NavigationMode.traditional;
     switch (mode) {
       case NavigationMode.traditional:
         return isEnabled && widget.useNativeKeyboard;
@@ -46,10 +48,12 @@ class _PinputState extends State<Pinput>
     }
   }
 
-  TextEditingController get _effectiveController => widget.controller ?? _controller!.value;
+  TextEditingController get _effectiveController =>
+      widget.controller ?? _controller!.value;
 
   @protected
-  FocusNode get effectiveFocusNode => widget.focusNode ?? (_focusNode ??= FocusNode());
+  FocusNode get effectiveFocusNode =>
+      widget.focusNode ?? (_focusNode ??= FocusNode());
 
   @protected
   bool get hasError => widget.forceErrorState || _validatorErrorText != null;
@@ -70,7 +74,8 @@ class _PinputState extends State<Pinput>
   @override
   void initState() {
     super.initState();
-    _gestureDetectorBuilder = _PinputSelectionGestureDetectorBuilder(state: this);
+    _gestureDetectorBuilder =
+        _PinputSelectionGestureDetectorBuilder(state: this);
     if (widget.controller == null) {
       _createLocalController();
       _recentControllerValue = TextEditingValue.empty;
@@ -88,7 +93,8 @@ class _PinputState extends State<Pinput>
   /// Android Autofill
   void _maybeInitSmartAuth() async {
     final isAndroid = UniversalPlatform.isAndroid;
-    final isAutofillEnabled = widget.androidSmsAutofillMethod != AndroidSmsAutofillMethod.none;
+    final isAutofillEnabled =
+        widget.androidSmsAutofillMethod != AndroidSmsAutofillMethod.none;
 
     if (isAndroid && isAutofillEnabled) {
       _smartAuth = SmartAuth();
@@ -98,14 +104,16 @@ class _PinputState extends State<Pinput>
   }
 
   void _maybePrintAppSignature() async {
-    if (widget.androidSmsAutofillMethod == AndroidSmsAutofillMethod.smsRetrieverApi) {
+    if (widget.androidSmsAutofillMethod ==
+        AndroidSmsAutofillMethod.smsRetrieverApi) {
       final res = await _smartAuth!.getAppSignature();
       debugPrint('Pinput: App Signature for SMS Retriever API Is: $res');
     }
   }
 
   void _listenForSmsCode() async {
-    final useUserConsentApi = widget.androidSmsAutofillMethod == AndroidSmsAutofillMethod.smsUserConsentApi;
+    final useUserConsentApi = widget.androidSmsAutofillMethod ==
+        AndroidSmsAutofillMethod.smsUserConsentApi;
     final res = await _smartAuth!.getSmsCode(
       useUserConsentApi: useUserConsentApi,
       matcher: widget.smsCodeMatcher,
@@ -121,7 +129,8 @@ class _PinputState extends State<Pinput>
   }
 
   void _handleTextEditingControllerChanges() {
-    final textChanged = _recentControllerValue.text != _effectiveController.value.text;
+    final textChanged =
+        _recentControllerValue.text != _effectiveController.value.text;
     _recentControllerValue = _effectiveController.value;
     if (textChanged) {
       _onChanged(pin);
@@ -189,7 +198,9 @@ class _PinputState extends State<Pinput>
 
   void _createLocalController([TextEditingValue? value]) {
     assert(_controller == null);
-    _controller = value == null ? RestorableTextEditingController() : RestorableTextEditingController.fromValue(value);
+    _controller = value == null
+        ? RestorableTextEditingController()
+        : RestorableTextEditingController.fromValue(value);
     _controller!.addListener(_handleTextEditingControllerChanges);
     if (!restorePending) {
       _registerController();
@@ -217,21 +228,35 @@ class _PinputState extends State<Pinput>
     TextSelection selection,
     SelectionChangedCause? cause,
   ) {
+    _effectiveController.selection =
+        TextSelection.collapsed(offset: pin.length);
+
     switch (Theme.of(context).platform) {
       case TargetPlatform.iOS:
       case TargetPlatform.macOS:
-        if (cause == SelectionChangedCause.longPress || cause == SelectionChangedCause.drag) {
-          _editableText?.bringIntoView(selection.extent);
-        }
-        return;
       case TargetPlatform.linux:
       case TargetPlatform.windows:
       case TargetPlatform.fuchsia:
       case TargetPlatform.android:
-        if (cause == SelectionChangedCause.drag) {
+        if (cause == SelectionChangedCause.longPress ||
+            cause == SelectionChangedCause.drag) {
           _editableText?.bringIntoView(selection.extent);
         }
-        return;
+        break;
+    }
+
+    switch (Theme.of(context).platform) {
+      case TargetPlatform.iOS:
+      case TargetPlatform.fuchsia:
+      case TargetPlatform.android:
+        break;
+      case TargetPlatform.macOS:
+      case TargetPlatform.linux:
+      case TargetPlatform.windows:
+        if (cause == SelectionChangedCause.drag) {
+          _editableText?.hideToolbar();
+        }
+        break;
     }
   }
 
@@ -294,7 +319,8 @@ class _PinputState extends State<Pinput>
         forcePressEnabled = false;
         textSelectionControls ??= cupertinoDesktopTextSelectionControls;
         handleDidGainAccessibilityFocus = () {
-          if (!effectiveFocusNode.hasFocus && effectiveFocusNode.canRequestFocus) {
+          if (!effectiveFocusNode.hasFocus &&
+              effectiveFocusNode.canRequestFocus) {
             effectiveFocusNode.requestFocus();
           }
         };
@@ -312,7 +338,8 @@ class _PinputState extends State<Pinput>
         forcePressEnabled = false;
         textSelectionControls ??= desktopTextSelectionControls;
         handleDidGainAccessibilityFocus = () {
-          if (!effectiveFocusNode.hasFocus && effectiveFocusNode.canRequestFocus) {
+          if (!effectiveFocusNode.hasFocus &&
+              effectiveFocusNode.canRequestFocus) {
             effectiveFocusNode.requestFocus();
           }
         };
@@ -325,6 +352,7 @@ class _PinputState extends State<Pinput>
       initialValue: _effectiveController.text,
       builder: (FormFieldState<String> field) {
         return TextFieldTapRegion(
+          onTapOutside: widget.onTapOutside,
           child: MouseRegion(
             cursor: _effectiveMouseCursor,
             onEnter: (PointerEnterEvent event) => _handleHover(true),
@@ -358,7 +386,10 @@ class _PinputState extends State<Pinput>
     );
   }
 
-  Widget _buildEditable(TextSelectionControls? textSelectionControls, FormFieldState<String> field) {
+  Widget _buildEditable(
+    TextSelectionControls? textSelectionControls,
+    FormFieldState<String> field,
+  ) {
     final formatters = <TextInputFormatter>[
       ...widget.inputFormatters,
       LengthLimitingTextInputFormatter(
@@ -371,6 +402,7 @@ class _PinputState extends State<Pinput>
       child: UnmanagedRestorationScope(
         bucket: bucket,
         child: EditableText(
+          key: editableTextKey,
           maxLines: 1,
           style: PinputConstants._hiddenTextStyle,
           onChanged: (value) {
@@ -387,10 +419,8 @@ class _PinputState extends State<Pinput>
           enableIMEPersonalizedLearning: false,
           textInputAction: widget.textInputAction,
           textCapitalization: widget.textCapitalization,
-          toolbarOptions: widget.toolbarOptions,
           selectionColor: Colors.transparent,
           keyboardType: widget.keyboardType,
-          textDirection: TextDirection.ltr,
           obscureText: widget.obscureText,
           onSubmitted: (s) {
             widget.onSubmitted?.call(s);
@@ -401,7 +431,6 @@ class _PinputState extends State<Pinput>
           textAlign: TextAlign.center,
           autofocus: widget.autofocus,
           inputFormatters: formatters,
-          key: editableTextKey,
           restorationId: 'pinput',
           clipBehavior: Clip.hardEdge,
           cursorColor: Colors.transparent,
@@ -412,20 +441,23 @@ class _PinputState extends State<Pinput>
           backgroundCursorColor: Colors.transparent,
           selectionHeightStyle: BoxHeightStyle.tight,
           enableSuggestions: widget.enableSuggestions,
-          onSelectionChanged: _handleSelectionChanged,
           contextMenuBuilder: widget.contextMenuBuilder,
           obscuringCharacter: widget.obscuringCharacter,
           onAppPrivateCommand: widget.onAppPrivateCommand,
+          onSelectionChanged: _handleSelectionChanged,
           onSelectionHandleTapped: _handleSelectionHandleTapped,
           readOnly: widget.readOnly || !isEnabled || !widget.useNativeKeyboard,
-          selectionControls: widget.toolbarEnabled ? textSelectionControls : null,
-          keyboardAppearance: widget.keyboardAppearance ?? Theme.of(context).brightness,
+          selectionControls:
+              widget.toolbarEnabled ? textSelectionControls : null,
+          keyboardAppearance:
+              widget.keyboardAppearance ?? Theme.of(context).brightness,
         ),
       ),
     );
   }
 
-  MouseCursor get _effectiveMouseCursor => MaterialStateProperty.resolveAs<MouseCursor>(
+  MouseCursor get _effectiveMouseCursor =>
+      MaterialStateProperty.resolveAs<MouseCursor>(
         widget.mouseCursor ?? MaterialStateMouseCursor.textable,
         <MaterialState>{
           if (!isEnabled) MaterialState.disabled,
@@ -437,7 +469,8 @@ class _PinputState extends State<Pinput>
 
   void _semanticsOnTap() {
     if (!_effectiveController.selection.isValid) {
-      _effectiveController.selection = TextSelection.collapsed(offset: _effectiveController.text.length);
+      _effectiveController.selection =
+          TextSelection.collapsed(offset: _effectiveController.text.length);
     }
     _requestKeyboard();
   }
@@ -460,7 +493,8 @@ class _PinputState extends State<Pinput>
           <Listenable>[effectiveFocusNode, _effectiveController],
         ),
         builder: (BuildContext context, Widget? child) {
-          final shouldHideErrorContent = widget.validator == null && widget.errorText == null;
+          final shouldHideErrorContent =
+              widget.validator == null && widget.errorText == null;
 
           if (shouldHideErrorContent) return onlyFields();
 
@@ -483,7 +517,8 @@ class _PinputState extends State<Pinput>
   @protected
   bool get hasFocus {
     final isLastPin = selectedIndex == widget.length;
-    return effectiveFocusNode.hasFocus || (!widget.useNativeKeyboard && !isLastPin);
+    return effectiveFocusNode.hasFocus ||
+        (!widget.useNativeKeyboard && !isLastPin);
   }
 
   @protected
@@ -501,7 +536,9 @@ class _PinputState extends State<Pinput>
           padding: const EdgeInsetsDirectional.only(start: 4, top: 8),
           child: Text(
             _errorText!,
-            style: widget.errorTextStyle ?? theme.textTheme.titleMedium?.copyWith(color: theme.colorScheme.error),
+            style: widget.errorTextStyle ??
+                theme.textTheme.titleMedium
+                    ?.copyWith(color: theme.colorScheme.error),
           ),
         );
       }
@@ -512,11 +549,13 @@ class _PinputState extends State<Pinput>
 
   // AutofillClient implementation start.
   @override
-  void autofill(TextEditingValue newEditingValue) => _editableText!.autofill(newEditingValue);
+  void autofill(TextEditingValue newEditingValue) =>
+      _editableText!.autofill(newEditingValue);
 
   @override
   TextInputConfiguration get textInputConfiguration {
-    final List<String>? autofillHints = widget.autofillHints?.toList(growable: false);
+    final List<String>? autofillHints =
+        widget.autofillHints?.toList(growable: false);
     final AutofillConfiguration autofillConfiguration = autofillHints != null
         ? AutofillConfiguration(
             uniqueIdentifier: autofillId,
@@ -526,6 +565,7 @@ class _PinputState extends State<Pinput>
           )
         : AutofillConfiguration.disabled;
 
-    return _editableText!.textInputConfiguration.copyWith(autofillConfiguration: autofillConfiguration);
+    return _editableText!.textInputConfiguration
+        .copyWith(autofillConfiguration: autofillConfiguration);
   }
 }