+++ lib/src/widgets/_pinput_selection_gesture_detector_builder.dart
@@ -11,7 +11,7 @@ class _PinputSelectionGestureDetectorBuilder
   @override
   void onForcePressStart(ForcePressDetails details) {
     super.onForcePressStart(details);
-    if (shouldShowSelectionToolbar) {
+    if (delegate.selectionEnabled && shouldShowSelectionToolbar) {
       editableText.showToolbar();
     }
   }